updated can communication to try and make the can2 work. Not working yet

+++ zephyr/modules/owntech_communication/zephyr/public_api/CanCommunication.cpp
@@ -51,7 +51,7 @@ bool CanCommunication::getCtrlEnable()
     return ctrl_enable;
 }
 
-float32_t CanCommunication::getCtrlReference()
+uint16_t CanCommunication::getCtrlReference()
 {
     return can_reference_value;
 }
@@ -77,7 +77,7 @@ void CanCommunication::setCtrlEnable(bool enable)
     ctrl_enable = enable;
 }
 
-void CanCommunication::setCtrlReference(float32_t reference)
+void CanCommunication::setCtrlReference(uint16_t reference)
 {
     can_reference_value = reference;
 }
